backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.31.0
httpx>=0.27.0
PyJWT>=2.8.0
openai>=1.0.0
boto3>=1.34.0
python-multipart>=0.0.9
python-docx>=1.1.0
pypdf>=4.0.0
qdrant-client>=1.7.0
neo4j>=5.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
cohere>=4.0
redis>=5.0.0
rq>=1.16.0
pgvector>=0.2.0
tencentcloud-sdk-python>=3.0.0
